erased read buffer

+++ pipex.h
@@ -11,7 +11,6 @@
 #define READ_START 0
 #define WRITE_END 1
 
-int ft_strlen(char *str);
-
+size_t ft_strlen(char *str);
 
 #endif
